[Journal] Starting off with Boilerplate Code
This commit will be the first of many, although I would like to use this special occasion to explain a little bit behind my motivation to create this CLI application, as well as to provide details on the things that are being committed.

First of all, Clockify is a SaaS application that is used at the HCB Organisation I work at, Institute Alterna. Although using Clockify is fairly easy to use with their web application, making use of the CLI to interact with the Clockify API seems fun and an interesting project to undertake.

During the first few lines and structure of the project as a whole, I felt like 13-year-old me who made his first Node.js application in the shape of a Discord bot, in companion with tutorials from YouTube and reading through documentation: a practise that I aim to uphold as I build this CLI utility.

This commit lays down the foundation to what will come next, importing modules, utilities, and packages neccesary for the development's success. I look forward to update you more on my progress as I continue to create this application.

Esteemed reviewer, please consider commits prefixed with [Journal] as complete, full-on explanations of the progress made since the last Journal or Checkpoint, while the committs prefixed with [Checkpoint] are intended to quickly update you in terms of the building bricks that are being place in order to construct a full-on function. I hope you have a good time reviewing my project!

- Sebs