[Journal] Initial Onboarding Process + Authentication & Process Exit (Centralised Error Handling, multiple utilities added)
Implemented an onboarding process that will be run when the user attempts to run any command of the utility for the first time: the goal is to obtain a valid API key and allow for proper command operation. Likewise, error handling has been put in place (and centralised on the src/utils/error.js file) in order to provide the user with a better experience when encountering unexpected errors. Whenever the program runs into an error, it will exit the process after displaying the error message. Additionally, error handling also covers exiting the application with (CTRL+C) in order to have a brief farewell message!

It is also important to mention that I have added multiple utilities to the src/utils directory, which will enable uniform reusability across the application. I hope there's no such thing as too much utility files! I checked out the GitHub CLI (gh) program which had a lot of separate utilities scripts, so I believe there's nothing wrong with that.

A major inspiration for me for this particular implementation came from my personal experience with the Heroku CLI. I definetly like the feeling of "unboxing" a brand new app, specially when it comes to the CLI, as it may be alien to some.

I'll continue to work on this throughout the week, before moving on to the actual operations with the API.

- Sebs