[Checkpoint] Added ESLint & Prettier Configuration for Code Quality and Formatting
I've added these two formatting tools in order to keep a uniform styling throughout the code I commit. I made it possible so that whoever clones the repository is able to access these settings and quickly make use of these formatting tools, which can be run with npm scripts as well.

- Added ESLint with Google style guide and Prettier integration.
- Configured scripts for linting and fixing code style in package.json.
- Updated auth.js, project.js, config.js, error.js, keyManager.js, onboarding.js, and userInfo.js for consistent code style.
- Introduced .prettierrc and .prettierignore for Prettier settings.
- Created .vscode/settings.json for editor configuration.
- Added eslint.config.js for ESLint rules and settings.