[Checkpoint] Added Debug utility with VSC, added API key import from argument, and uniformalised error handling
I've set up a debugger with VSC through a JavaScript Debug Terminal, which will allow me to make use of breakpoints much more effectively. Most importantly, I added support to import an API key through using an argument in the clockify auth login command.

Likewise, I've improved error handling through the use of sytnax in error.js for it to be the default export. Last but not least, I also improved user info retrieval by disposing of a if/else's else in order to make the code much clearer