[Checkpoint] Added API modals for projects, time entries, and workspace data.
I've set up multiple modules which will provide API calls to fulfill the upcoming timer.js command, which calls for projects, time entries, and workspace data. I've also relocated the getUserInfo() method file to be located in the api directory, so it is more tidy and organised.

Finally, I imported some workspace data onto the clockify auth status command, so it is well adorned with the organisation name as well. It won't be long until one is able to start a timer directly from the CLI. Looking forward to finishing it before the end of the week.