[Journal] Timer Command + Uniformation
Starting and stopping timers within the CLI is the minimum viable product (in my opinion) of this project. After extensive trial and error, I was able to successfully do this by making use of API payloads which I also uniformed through an apiHeaders.js file which will ensure all headers are being passed to functions consistently.

Enabling the new timer command, which allows the user to start, stop, and view the current status of timers was possibile by modifying how more than 4 datatypes are handles by the app, being these projects, timeEntries, userInfo, and workspace, which all work in coordination to ensure data is formatted correctly. It is slightly confusing on how these data types interconnect but it seems to also be a way to optimise the app by only calling certain data types on later commands to be developed.

Other than that, I've also updated the semantic version of the CLI to comply with the modern standard. Up next, I aim to fix a bug where the "project" argument in the timer command won't render correctly. I assume I must find the project through an API utility soon to be placed in the projects.js file.