[Checkpoint] Add delete command for timers and enchance timer duration calculation
This commit introduces the use of complex one-liner logic in order to parse information without repeating oneself, with the clockify timer delete command as a first instance of this. Likewise, the use of Promise.all has been put in place in order to experiment with potential performance improvements. Finally, timerDuration.js will now compute a timer length if a end-of-timer timestamp is provided optionally.